员工数据修改以及新增员工Bug修复

+++ sky-server/src/main/java/com/sky/controller/admin/EmployeeController.java
@@ -125,7 +125,7 @@ public class EmployeeController {
      */
     @GetMapping("/{id}")
     @ApiOperation(value = "根据id查询员工")
-    public Result<Employee> queryById(@PathVariable Integer id) {
+    public Result<Employee> queryById(@PathVariable Long id) {
         Employee employee = employeeService.queryById(id);
         return Result.success(employee);
     }
